#67 removed page captions and started moving resume to be more aligned with experiences
