--- src/data/Resume/ResumeSkillsData.ts
@@ -1,44 +1,0 @@
-interface SkillInformation {
-  category: string;
-  tools: string[];
-}
-
-export const skills: SkillInformation[] = [
-  {
-    category: 'Programming Languages',
-    tools: [
-      'JavaScript/TypeScript',
-      'HTML/CSS',
-      'Java',
-      ' Python',
-      'C#',
-      'C/C++',
-      'SQL',
-    ],
-  },
-  {
-    category: 'Frameworks/Libraries',
-    tools: [
-      'React',
-      'Material UI',
-      'Express',
-      'Node.js',
-      'JUnit',
-      'Next.js',
-      'Gatsby',
-    ],
-  },
-  {
-    category: 'Applications/Tools',
-    tools: [
-      'Git',
-      'GitHub',
-      'Figma',
-      'Illustrator',
-      'InDesign',
-      'Photoshop',
-      'Unity',
-    ],
-  },
-];
-
