+++ src/pages/code.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import PageHeading from '../components/PageHeading';
-import { Typography, Divider } from '@mui/material';
-import ComingSoon from '../components/ComingSoon';
+import { Divider } from '@mui/material';
 import ProjectCard from '../components/CodePage/ProjectCard';
 import { projects } from '../data/ProgrammingProjects/ProjectsData';
 import { GridMainContainer } from '../components/Layout/GridFormatting';
@@ -10,11 +9,6 @@ export default function CodePage() {
   return (
     <>
       <PageHeading title='Code' />
-      <Typography variant='subtitle1'>
-        {
-          "Here are some projects I've worked on over the years. I particularly like working on frontend related things."
-        }
-      </Typography>
       <Divider />
       <GridMainContainer>
         {projects.map((project) => ProjectCard(project))}
